+++ CHANGELOG.md
@@ -7,6 +7,9 @@ For older changes see the [archived Singularity change log](https://github.com/a
 
 ## Changes for v1.4.x
 
+- Label process for starter binary of interactive containers with image filename,
+  for example: `Apptainer runtime parent: example.sif`.
+
 ## v1.3.2 - \[2024-05-28\]
 
 ### Security fix